+++ lib/stroop/cli.rb
@@ -5,29 +5,32 @@ module Stroop
   class CLI < Thor
     DEFAULT_SIZE = 5.freeze
 
+    class_option :seed, type: :numeric, aliases: "-s", description: "The random seed to generate the color words"
+
     desc 'neutral COLSxROWS', 'prints neutral color words in COLS columns and ROWS rows'
     def neutral(size = '10x5')
-      print(size: size, mode: Set::NEUTRAL)
+      print(size: size, mode: Set::NEUTRAL, seed: options[:seed])
     end
 
     desc 'congruent COLSxROWS', 'prints congruent color words in COLS columns and ROWS rows'
     def congruent(size = '10x5')
-      print(size: size, mode: Set::CONGRUENT)
+      print(size: size, mode: Set::CONGRUENT, seed: options[:seed])
     end
 
     desc 'incongruent COLSxROWS', 'prints incongruent color words in COLS columns and ROWS rows'
     def incongruent(size = '10x5')
-      print(size: size, mode: Set::INCONGRUENT)
+      print(size: size, mode: Set::INCONGRUENT, seed: options[:seed])
     end
 
     private
 
-    def print(size:, mode:)
+    def print(size:, mode:, seed:)
       rows, columns = dimensions(size)
       rows          = apply_default_size(rows)
       columns       = apply_default_size(columns)
 
-      puts Set.new(rows: rows, columns: columns, mode: mode).to_s
+      set = Stroop::Set.new(rows: rows, columns: columns, mode: mode, seed: seed)
+      puts set, "🌱 Generated with seed: #{set.seed}"
     end
 
     def dimensions(size)